requirements.txt
django
djangorestframework
psycopg2-binary
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.9
djangorestframework==3.12.4
ipython==7.29.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
sqlparse==0.4.2
traitlets==5.1.1
wcwidth==0.2.5
